+++ core/execution/engine_snapshot.go
@@ -198,7 +198,7 @@ func (e *Engine) restoreMarket(ctx context.Context, em *types.ExecMarket) (*futu
 	}
 
 	nextMTM := time.Unix(0, em.NextMTM)
-	nextIndexPriceCalc := time.Unix(0, em.NextIndexPriceCalc)
+	nextInternalCompositePriceCalc := time.Unix(0, em.NextInternalCompositePriceCalc)
 
 	// create market auction state
 	e.log.Info("restoring market", logging.String("id", em.Market.ID))
@@ -244,7 +244,7 @@ func (e *Engine) restoreMarket(ctx context.Context, em *types.ExecMarket) (*futu
 	}
 	// ensure this is set correctly
 	mkt.SetNextMTM(nextMTM)
-	mkt.SetNextIndexPriceCalc(nextIndexPriceCalc)
+	mkt.SetNextInternalCompositePriceCalc(nextInternalCompositePriceCalc)
 
 	e.publishNewMarketInfos(ctx, mkt.GetMarketData(), *mkt.Mkt())
 	return mkt, nil