+++ core/integration/setup_test.go
@@ -439,8 +439,8 @@ func (e *executionTestSetup) registerNetParamsCallbacks() error {
 			Watcher: execsetup.executionEngine.OnMarkPriceUpdateMaximumFrequency,
 		},
 		netparams.WatchParam{
-			Param:   netparams.IndexPriceUpdateFrequency,
-			Watcher: execsetup.executionEngine.OnIndexPriceUpdateFrequency,
+			Param:   netparams.InternalCompositePriceUpdateFrequency,
+			Watcher: execsetup.executionEngine.OnInternalCompositePriceUpdateFrequency,
 		},
 		netparams.WatchParam{
 			Param:   netparams.SpamProtectionMaxStopOrdersPerMarket,